removed CacheData interfaces
